Fine-grained UI updates
Gets rid of `MainFrameRepaint` in favor of specific update functions
that tell the UI exactly what changed.

This improves the efficiency of various handlers. Also fixes problems
with mined transactions not showing up until restart.

The following notifications were added:

- `NotifyBlocksChanged`: Block chain changed
- `NotifyKeyStoreStatusChanged`: Wallet status (encrypted, locked)
changed.
- `NotifyAddressBookChanged`: Address book entry changed.
- `NotifyTransactionChanged`: Wallet transaction added, removed or
updated.
- `NotifyNumConnectionsChanged`: Number of connections changed.
- `NotifyAlertChanged`: New, updated or cancelled alert. As this finally
makes it possible for the UI to know when a new alert arrived, it can be
shown as OS notification.

These notifications could also be useful for RPC clients. However,
currently, they are ignored in bitcoind (in noui.cpp).

Also brings back polling with timer for numBlocks in ClientModel. This
value updates so frequently during initial download that the number of
signals clogs the UI thread and causes heavy CPU usage. And after
initial block download, the value changes so rarely that a delay of half
a second until the UI updates is unnoticable.