move class HelpMessageBox to guiutil.cpp/.h , etc.
- Add button to show Version command-line options (in RPC Console ->
Information)
- Resize Debug window a little to allow for a non-breaking display of
the welcome message with non-english translation

- Show command line options as dialog when opened from debug window
On Linux/Mac the command-line options were printed to stderr when the
button
was pressed in the debug window, resulting in confusion. This is fixed
in this commit by adding a separate method.

- Cross-platform "Open debug logfile"
The option to open the debug logfile from the debug window was
implemented only for
windows. By using `QDesktopServices::openUrl` it now works on any
platform.